update user account type in firestore

+++ stores/auth.ts
@@ -1,6 +1,6 @@
 import { defineStore } from 'pinia';
 import { createUserWithEmailAndPassword, signInWithEmailAndPassword, signOut } from '@firebase/auth';
-import { doc, setDoc, getDoc, collection, onSnapshot } from '@firebase/firestore';
+import { doc, updateDoc, getDoc, collection, onSnapshot } from '@firebase/firestore';
 import type { User } from 'firebase/auth';
 import { AccountType, type StaffData } from '@/types/auth';
 
@@ -64,20 +64,51 @@ export const useAuthStore = defineStore('users', {
       }
     },
 
-    async fetchCurrentUser(id:string) {
-      const nuxtApp = useNuxtApp()
-      try{
+    async fetchCurrentUser(id: string) {
+      const nuxtApp = useNuxtApp();
+      try {
         if (!id) {
-          throw new Error("Invalid user ID")
+          throw new Error("Invalid user ID");
         }
-        const userDocRef = doc(nuxtApp.$firestore, 'users',id)
-        const userDocSnapshot = await getDoc(userDocRef)
-        this.currentUser = userDocSnapshot.data()
-        localStorage.setItem('currentUser', JSON.stringify(this.currentUser))
-        return this.currentUser
+        const userDocRef = doc(nuxtApp.$firestore, 'users', id);
+        const userDocSnapshot = await getDoc(userDocRef);
+    
+        if (userDocSnapshot.exists()) {
+          const userData = userDocSnapshot.data();
+          this.currentUser = {
+            ...userData,
+            emailVerified: false,
+            isAnonymous: false,
+            metadata: {},
+            id,
+          } as User & {
+            password: string;
+            accountType: AccountType;
+            id: string;
+            adminName?: string;
+            imageUrl?: string;
+            adminId?: string;
+          };
+          localStorage.setItem('currentUser', JSON.stringify(this.currentUser));
+        } else {
+          throw new Error('User document does not exist.');
+        }
+        return this.currentUser;
+      } catch (error) {
+        console.error('Error fetching current user:', error);
       }
-      catch(error){
-        console.log(error)
+    },
+
+    async updateAccountType(userId: string, newAccountType: AccountType) {
+      const nuxtApp = useNuxtApp();
+      try {
+        const userDocRef = doc(nuxtApp.$firestore, 'users', userId);
+        await updateDoc(userDocRef, {
+          accountType: newAccountType,
+        });
+        console.log('Account type updated successfully.');
+      } catch (error) {
+        console.error('Error updating account type:', error);
       }
     },
 
@@ -90,7 +121,7 @@ export const useAuthStore = defineStore('users', {
           let userData = doc.data();
           userData.id = doc.id;
           if (userData.adminId === this.currentUser?.id) {
-            this.staffList.unshift(userData);
+            this.staffList.unshift(userData as StaffData);
           }
         });
       });
